requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.0
pyyaml==6.0.1
jinja2==3.1.4
httpx==0.27.0
pytest==8.2.0
pytest-asyncio==0.23.6
